opti, bitwise, clean, cos tables

+++ srcs/angles.c
@@ -6,20 +6,20 @@
 /*   By: gbianco <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/08/02 14:39:40 by gbianco           #+#    #+#             */
-/*   Updated: 2020/08/02 14:39:42 by gbianco          ###   ########.fr       */
+/*   Updated: 2020/08/03 22:18:32 by gbianco          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "wolf3d.h"
 
-void	protected_dif(double *angle, double dif)
+void	protected_dif(float *angle, float dif)
 {
 	*angle -= dif;
 	if (*angle < 0)
 		*angle = RADCIRC - -*angle;
 }
 
-void	protected_sum(double *angle, double sum)
+void	protected_sum(float *angle, float sum)
 {
 	*angle += sum;
 	if (*angle > RADCIRC)
